requirements.txt
Appium-Python-Client==3.0.0
pytest==7.4.2
selenium==4.12.0
